+++ frontend/src/services/AuthService.ts
@@ -3,7 +3,7 @@ import type { CognitoUserSession } from 'amazon-cognito-identity-js';
 import { Auth } from 'aws-amplify';
 import { readonly, ref } from 'vue';
 import { EnvironmentSettings } from '@/services/EnvironmentSettings';
-import { currentSelectedApplication } from './ApplicationState';
+import { currentSelectedApplication } from '@/store/ApplicationState';
 
 const FAM_LOGIN_USER = 'famLoginUser';
 
@@ -17,9 +17,9 @@ export interface FamLoginUser {
 const state = ref({
     famLoginUser: localStorage.getItem(FAM_LOGIN_USER)
         ? (JSON.parse(localStorage.getItem(FAM_LOGIN_USER) as string) as
-              | FamLoginUser
-              | undefined
-              | null)
+            | FamLoginUser
+            | undefined
+            | null)
         : undefined,
 });
