+++ frontend/src/store/ApplicationState.ts
